+++ src/OdometryCollector.cpp
@@ -9,11 +9,11 @@ Robot::OdometryCollector::OdometryCollector() : m_initial(0, 0, 0), m_pose(0, 0,
 
 Robot::OdometryCollector::OdometryCollector(Pose2D initial) : m_initial(initial), m_pose(initial) { }
 
-Robot::OdometryCollector::OdometryCollector(double x, double y, double theta) : m_initial(x, y, theta), m_pose(x, y, theta) { }
+Robot::OdometryCollector::OdometryCollector(float x, float y, float theta) : m_initial(x, y, theta), m_pose(x, y, theta) { }
 
 void Robot::OdometryCollector::odoTranslate(Pose2D offset) {
-    double dst = hypot(offset.X(), offset.Y());
-    double angle = atan2(offset.Y(), offset.X());
+    float dst = hypot(offset.X(), offset.Y());
+    float angle = atan2(offset.Y(), offset.X());
     m_pose.setX(m_pose.X() + (cos(m_pose.Theta() + angle) * dst));
     m_pose.setY(m_pose.Y() + (sin(m_pose.Theta() + angle + offset.Theta()) * dst));
     m_pose.setTheta(m_pose.Theta() + offset.Theta());
